Rearrange, simplify
