+++ daskms/testing.py
@@ -19,22 +19,3 @@ def mark_in_pytest(in_pytest=True):
         raise TypeError('in_pytest %s is not a boolean' % in_pytest)
 
     __pytest_run_marker__['in_pytest'] = in_pytest
-
-
-def requires(msg, *import_errors):
-    if any(isinstance(e, ImportError) for e in import_errors):
-        def decorator(fn):
-            def wrapper(*args, **kwargs):
-                if in_pytest():
-                    import pytest
-                    pytest.skip(msg)
-                else:
-                    raise ImportError(msg)
-
-            return wrapper
-    else:
-        def decorator(fn):
-            print(fn)
-            return fn
-
-    return decorator
